+++ src/module/data/item/armor-data.mjs
