+++ src/module/data/item/weapon-data.mjs
